requirements.txt
Django>=5.2.3,<5.3.0
djangorestframework>=3.14.0,<3.15.0
djangorestframework-simplejwt>=5.3.0,<5.4.0
django-cors-headers>=4.3.1,<4.4.0
drf-yasg>=1.21.7,<1.22.0
# Pillow>=10.1.0,<10.2.0
python-dotenv>=1.0.0,<1.1.0
dj-database-url>=2.1.0,<2.2.0
mysqlclient>=2.2.0,<2.3.0
gunicorn>=21.2.0,<21.3.0
whitenoise>=6.6.0,<6.7.0
